[runtime] Split `Blob::read_at` into `read_at` and `read_at_buf` (#3041)

+++ stream/src/utils/codec.rs
@@ -55,15 +55,15 @@ pub async fn send_frame<S: Sink>(
 /// stream is closed.
 pub async fn recv_frame<T: Stream>(stream: &mut T, max_message_size: u32) -> Result<IoBufs, Error> {
     let (len, skip) = recv_length(stream).await?;
-    if len > max_message_size {
-        return Err(Error::RecvTooLarge(len as usize));
+    if len > max_message_size as usize {
+        return Err(Error::RecvTooLarge(len));
     }
 
     stream
-        .recv(skip as u64 + len as u64)
+        .recv(skip + len)
         .await
         .map(|mut bufs| {
-            bufs.advance(skip as usize);
+            bufs.advance(skip);
             bufs
         })
         .map_err(Error::RecvFailed)
@@ -73,15 +73,15 @@ pub async fn recv_frame<T: Stream>(stream: &mut T, max_message_size: u32) -> Res
 /// Returns (payload_len, bytes_to_skip) where bytes_to_skip is:
 /// - varint_len if decoded from peek buffer (bytes not yet consumed)
 /// - 0 if decoded via recv (bytes already consumed)
-async fn recv_length<T: Stream>(stream: &mut T) -> Result<(u32, u32), Error> {
+async fn recv_length<T: Stream>(stream: &mut T) -> Result<(usize, usize), Error> {
     let mut decoder = Decoder::<u32>::new();
 
     // Fast path: decode from peek buffer without blocking
     let peeked = {
-        let peeked = stream.peek(MAX_U32_VARINT_SIZE as u64);
+        let peeked = stream.peek(MAX_U32_VARINT_SIZE);
         for (i, byte) in peeked.iter().enumerate() {
             match decoder.feed(*byte) {
-                Ok(Some(len)) => return Ok((len, i as u32 + 1)),
+                Ok(Some(len)) => return Ok((len as usize, i + 1)),
                 Ok(None) => continue,
                 Err(_) => return Err(Error::InvalidVarint),
             }
@@ -90,15 +90,12 @@ async fn recv_length<T: Stream>(stream: &mut T) -> Result<(u32, u32), Error> {
     };
 
     // Slow path: fetch bytes one at a time (skipping already-decoded peek bytes)
-    let mut buf = stream
-        .recv(peeked as u64 + 1)
-        .await
-        .map_err(Error::RecvFailed)?;
+    let mut buf = stream.recv(peeked + 1).await.map_err(Error::RecvFailed)?;
     buf.advance(peeked);
 
     loop {
         match decoder.feed(buf.get_u8()) {
-            Ok(Some(len)) => return Ok((len, 0)),
+            Ok(Some(len)) => return Ok((len as usize, 0)),
             Ok(None) => {}
             Err(_) => return Err(Error::InvalidVarint),
         }
@@ -175,7 +172,7 @@ mod tests {
             // 1024 (MAX_MESSAGE_SIZE) encodes as varint: [0x80, 0x08] (2 bytes)
             let read = stream.recv(2).await.unwrap();
             assert_eq!(read.coalesce(), &[0x80, 0x08]); // 1024 as varint
-            let read = stream.recv(MAX_MESSAGE_SIZE as u64).await.unwrap();
+            let read = stream.recv(MAX_MESSAGE_SIZE as usize).await.unwrap();
             assert_eq!(read.coalesce(), buf);
         });
     }
